refactor(ui): Clean up the refactor
Add an index file for barrel exporting all utilities, switch consumers to
use this barrel export, and delete the old `-create-run-utils.ts` file.
