+++ ui/src/routes/organizations/$orgId/products/$productId/repositories/$repoId/_repo-layout/create-run/index.tsx
@@ -70,7 +70,7 @@ import {
   defaultValues,
   flattenErrors,
   formValuesToPayload,
-} from './-create-run-utils';
+} from './-components';
 
 const CreateRunPage = () => {
   const navigate = useNavigate();
